do not override the KEYRING package
This step complicates the install at no real benefit in
security. Sure, we have a more direct trust path by controlling the
keyring package, but if we can't trust the keyring that's shipped with
debootstrap, we can't trust trust debootstrap to check a keyring
that *we* would provide either.

Furthermore, this procedure is exclusively used on Hetzner Cloud
machines, which *do* ship a sane keyring file, last I checked. Short
of auditing the entire machine, of which shipping a keyring is just a
tiny step, I don't think this is worth it.