expand puppet bootstrap script (#32914)
We now disable the puppet service *before* the package is installed,
so that it only runs under our command. (If that doesn't work, next
thing to try is a policy-rc.d thing.)

We also inject a hardcoded version of the Puppet CA. The rationale
behind this is that it reduces the number of back-and-forth between
the server and the client to the minimum, which is for the client to
ask for a cert (the CSR) and for the server to sign it.

The CA is also created *before* Puppet is ran so that we remove that
MITM vector altogether. There *is* still such a vector in the
installer itself, but that's outside our threat model in this specific
instance: if someone can backdoor the installer, the can do nasty
stuff outside of this script anyways.

Finally, we include more verbose directives so the user has a better
idea of what to do next.

This assumes the existance of a signing tool on the server which I
haven't created yet, that's the next step.