improve hostname check to cover the `hostname`=`hostname -f` case
It would have been possible, in the previous incantation, to have a
"hostname" that is the fqdn (and equal to `hostname -f`), so our check
wouldn't have checked /etc/hosts was correct.

This new check ensures that `hostname` is the short version and that
it loops around to get the right hostname in `hostname -f`.