append basic DNS configuration in bootstrap if missing (#32914)
This is critical to make Puppet bootstrap properly. If the domain
search is not setup correctly, Puppet will not find the `puppet`
domain and will fail to bootstrap itself.

This step should be done by the system installer, but it's possible it
failed at its job, either because it's buggy, incomplete, or that it
has a different set of assumptions than we do.

By appending the `domain` line, we make sure `puppet` will resolve to
`puppet.torproject.org` which should bootstrap Puppet. We also add a
dummy nameserver there (Google, for a lack of a better, well-known,
and public alternative) just in case `resolv.conf` really *is* badly
configured.

That configuration should be overriden by Puppet on the first run,
when the local resolver is installed.