give an idea of how to fix the hostname -f problem (#32914)
We don't try to fix /etc/hosts, because we don't have a fool-proof way
to determine the primary address of this machine which should be added
there.

We previously assumed the installer would do this for us, and we still
do. But at least we provide the operator a better idea of how to fix
this problem.