make swap larger
If we have (say) 64GB of ram, having a 32GB tmpfs and 1GB swap makes
no sense at all: users will be able to eat half of the core memory and
the kernel will have no recourse but to give it to them.

We should have swap at *least* as big as the tmpfs, which is by
default 50% of RAM. But while we're at it, just follow the
traditionnal convention of matching the core memory size. This will
give us spare room on the VG in case of trouble, and allows crash
dumps, along with giving us the room we need (and more) for the tmpfs,
and other such niceties.

In the gnt-fsn cluster we have 1TB of space there. We can spare the
expense.