set the list of base packages in our config instead of the wiki
We use an include list instead of an exclude list, because it seems
the latter has gotten as long as the former. It will also make it
clearer what we're actually installing in the base system, and this,
in turn, can be reused by other installers than the "robot" installer.