don't override module loading, it breaks main execution
Instead, we load the modules in the fabfile, which then exposes it
when we run fab from the top-level.

We still haven't quite figured out how to run the modules themselves
directly as executables, but that's typical Python loading horror that
I can't quite figure out right now. It works with `-m`, so document
that in a README for now.