do *not* adopt instance by default
This step is tricky: the disk order might be wrong, the disks might
not be fully coherent (because suspend is not implemented yet), the
operator doesn't have time to fire up the console to find problems,
and a multitude of other things can go wrong.

So use a sane default. We show the operator how to do the step
themselves (in verbose mode) anyways.