reverse argument order on most tasks, to put instance first
This will make it easier to run commands on multiple servers, as the
"instance" we operate on can only be the first argument.

In some cases, like `ganeti.stop`, it makes little sense from a coding
perspective because we don't use the context at all apart from its
config, but from a *user* perspective it's much more intuitive. It
turns this:

    fab -H fsn-node-01.torproject.org stop test.torproject.org

into this:

    fab -H test.torproject.org stop --master-host=fsn-node-01.torproject.org

or even this:

    fab -H test.torproject.org stop

and this:

    fab -H test.torproject.org,test2.torproject.org stop

This is a major change that hasn't been tested yet, but I'm confident
all cases were covered, at least internally.

This makes everything consistent with the way libvirt-import
works. The following diff is applied in the wiki docs, to give an
idea:

modified   tsa/howto/ganeti.mdwn
@@ -249,7 +249,7 @@ Import procedure:

  4. renumber the host:

-        ./ganeti -v -H $SPARE_NODE renumber-instance $INSTANCE
+        ./ganeti -v -H $INSTANCE renumber-instance --ganeti-node $SPARE_NODE

     TODO: this does not deal with `/etc/hosts` on the INSTANCE, which
     should be done by hand.
@@ -267,7 +267,7 @@ Import procedure:
  8. redo migration as in step 3 and 4:

         ./ganeti -v -H $INSTANCE libvirt-import  --ganeti-node $SPARE_NODE --libvirt-host $KVM_HOST --adopt
-        ./ganeti -v -H $SPARE_NODE renumber-instance $INSTANCE
+        ./ganeti -v -H $INSTANCE renumber-instance --ganeti-node $SPARE_NODE

  9. final test procedure

@@ -288,7 +288,7 @@ Import procedure:

  12. retire old instance (only a tiny part of [[retire-a-host]]):

-        ./retire -H kvmX.torproject.org retire-instance -i test.torproject.org
+        ./retire -H $INSTANCE retire-instance --parent-host $KVM_HOST

      TODO: the above is backwards - we should be able to do `retire -H
      test.torproject.org revoke-puppet retire-instance` instead...