secure SSH key fingerprint verification routines in Fabric
I found out that Fabric does not enforce SSH key fingerprints at all
by default, overriding the default (and proper) Paramiko policy in the
Connection object. Provide a better Connection object that overrides
that behavior and enforce it in the fabfile.

That class override fails because of a limitation in the way
hierarchies are setup in Fabric right now, so also warn about this.

What a mess!