rewire how we process regular expressions in ensure_line
We previously were unconditionnally passing DOTALL, which is tricky
because all of a sudden the dot metacharacter (`.`)
matches *everything*, including newline.

It makes for awkward combinations with the ensure_newline
parameter (which is default) and was causing in the hosts file
management unit tests (and its upcoming ipv6 implementation).

This makes it clearer and easier to define what ensure_newline means:
it applies only when appending a newline, and not replacing a new
one. The logic behind this is that, by default, we do NOT match the
newline character in the match pattern, so we would append duplicate
newlines in our search patterns.

This was also a problem in ssh keys replacement, for what that's
worth, and required serious regex messing around in that because of
the DOTALL it uses. That still uses DOTALL because it does match
multiple lines, but at least the general case is more sensical.