properly pass ipv4 and ipv6 addresses to rewrite_hosts
Previous call was passing a tuple to the function instead of
bytes. I'm surprised it worked at all, but that's Python for ya.

This, incidentally, starts adding IPv6 addresses to /etc/hosts, but
that seems reasonable: otherwise bacula still crashes during the
migration process and we don't like that.

See this comment for details, again:

https://trac.torproject.org/projects/tor/ticket/33732#comment:1