do not return in context manager
First of all, it doesn't really make sense: the context manager does
not "return" a usable value in any way, it "yields" one, and only one.

Second, it breaks exception handling. In this case, there was a bug in
the rewrite_hosts call in ganeti and it was being shadowed by the
return call. It's not clear to me why that was happening, but it
definitely made debugging really hard.