make shutdown and wait procedure supports halting the machine
Reboots are nice, but sometimes we want to completely turn off the
server. This makes the wait procedure support both processes.

This could also have been implemented in a separate function, but it
seemed that it would have duplicated too much code for this.

The function name was changed to make sure the callers adapt to the
new argument (kind) inserted as first.