do not append to file to rewrite the interfaces file
This would create a corrupt interfaces file if the backup was already
present, as we would previously *append* to the file. It's unclear why
we did this in the first place: we probably reused the "append"
procedure that was used before ensure_line() was created.

It's definitely not appropriate in this case: we rewrite the entire
file, as the function name indicates.