mark all the places where we do not do proper error handling with XXX
As it turns out, a @task, by default, will silently exit when an
exception occurs, by design:

http://www.pyinvoke.org/faq.html#everything-just-exits-silently-after-i-run-a-command

... unless warn=True. So we should review *everywhere* we call
context.run() and make sure we handle errors from there directly.

It's still unclear if this can be done by wrapping the call in a
try/except, or if we absolutely need warn=True. Some more prototyping
is necessary. But we definitely need to review this because the hacks
in 0d8acd1 (use wrapper functions for tasks) is really annoying and
should not be necessary if we do things right (and we should).

I tested the behavior this prototype:

    #!/usr/bin/python3

    from invoke import task

    @task
    def fun(con):
        print("this is fun!")
        print('running false')
        con.run('false')
        print('ran false')

The "ran false" string is never printed. This is, roughly, the
equivalent of `set -e` in a shell.