handle hosts with a crypto initrd that prompt for a password
The previous version of this code would sent the string "uptime" to
the initrd password prompt, which would totally hang it, probably
because we were not using `pty=True`. Now we not only do that
properly, but we also expect mandos to just reboot the node.

Alternatively, we could just let the user enter the passphrase here,
but the hope is that we have mandos everywhere for now.