allow caller to pass network name to adopt
We ended up not using the network listing because i don't know how I
would make the default autodetected properly. I considered making
network_name=True being the default, but then that would make Fabric
treat the argument as a on/off switch and we wouldn't be allowed to
provide a string. If the default is "None", then I'm not sure how
Fabric behaves at all.

Better to just use a proper default since this code will not live
forever anyways.