hardcode a default `root` user to connect to servers
Our Fabric setup assumes the user has its ssh keys configured in the
remote servers's root account everywhere. But we were assuming this
was configured in the user's `~/.ssh/config` file yet there was no
documentation on how to do so anywhere. Besides, it's better if this
just works correctly outside the box, so define the user in
`fabric.py` next to the `fabfile.py`.

This can still be overriden in ~/.ssh/config, so we insist on using
the `root` account in certain narrow cases. We were already doing so
in the ganeti removal step, but now we also explicitely do it in
revoke-puppet.