add the `host.id` command to debug connexion problems
I was constantly battling fabric to figure out if it was using my
configuration or not. Turns out that if you just pass a command to
`fab`, it does *not* load the project-specific configuration file,
because it doesn't need to load the `fabfile.py`. So having this
command is actually useful, even though it is trivial.