put ipv4 blocks after ipv6 in network configuration
Otherwise the accept_ra might not take effect soon enough and the
network interface might have a IPv6 gateway before it's
configured. We've seen this in tb-tester-01 (in the gnt-chi cluster),
where networking.service would fail with:

      RTNETLINK answers: File exists

... when trying to create the gateway. Turns out the IPv4
gateway (e.g. 10.0.0.1) would show up as a link-local
gateway (e.g. fe80::10:0:0:1) before ifupdown had time to set the
sysctl.

This will not fix existing configurations but will hopefully fix next
boxes.

An alternative would be to just not hardcode the IP address and rely
on RA, but we don't want to change policy just yet.