merge the LDAP bootstrap inside puppet
This will make it slightly easier because we run only one command
which will walk us through the next steps. It's kind of annoying that
we still need the docs on how to start `ldapvi` properly here though,
and the explanations of all the fields... Maybe that should be in
puppet-boostrap as well?

The idea is that, in the long term, we might not need to bootstrap
LDAP at all, and just keep host metadata in Puppet as the only source
of truth...