show utc timestamps to reduce confusion
The rationale here is that previously we didn't have timezone
information in the output at all. This was showing local time, but
that was confusing if we would copy-paste this to other users or
compare it with time on servers (which is typically UTC).

We could have done something crazy to guess the local timezone
instead (like https://stackoverflow.com/a/39079819/1174784) but this
seems simpler for now.

It can also be changed somewhat easily, now that we have the
`timezone.utc` object lying about.