handle-abuse: correctly fail when we do not unsubscribe users
I'm using this in notmuch-emacs, and it doesn't show the output of the
command unless it fails with an actual non-zero exit code. Somehow I
assumed this was already the case and only noticed because I was
looking at logs at the same time and didn't see the unsubscribe mail
coming out.