handle the case where the user didn't confirm usage soon enough
The Yubikey (for example) touch policy times out after 15 seconds,
which is annoying for users that have that enabled: they walk away
from the reboot and work on something else, then this silently prompts
them for the touch, and they miss it, and then fabric goes down in
flame and fails to continue.

Instead of just retrying three times, we just completely stop and
prompt the user if we have failed. We also send a bell to raise
urgency hints on windows.

If the user was paying attention and confirmed in time, this is a noop
and won't affect that use case.