handle all SSH exceptions specially
We can't seem to isolate the exception that is triggered by a yubikey
failure. The `saved_exception` is None and we never get access to the
actual exception we see in the debug output.

So let's just treat them all as a temporary failure and allow the
operator to interact.