fallback to BOOT_DEVICE of BOOT_PARTITION is missing
Commit 068503ce introduced a bug in the Hetzner Robot installer where
/target/boot was always empty. I suspect this is because the patch
was made for the installer to work on the gnt-chi cluster, which uses
hardware RAID. Maybe what happens is that FAI gives us two different
variables depending on the configuration.

The commit that introduced this bug is:

    commit 068503ce4ef843660d9ce3990c0508927bc0ee60
    Author:     Antoine Beaupré <anarcat@debian.org>
    Date: Mon Nov 2 16:18:06 2020 -0500

    /boot is the boot *partition*, not device

    This should fix the empty /boot problem in future installs.

Now we still default to the BOOT_PARTITION, but fallback if the
variable isn't defined, which seems to be the case in our current
tests with the gnt-fsn-NVMe disk configuration.