working dal-rescue disk configuration
This has been tested on the dev server and *seems* to work, or at
least it compiles. The previous p=foo configuration didn't work
because those were introduced only in bookworm.

We also lower significantly the disk usage because we're quite cramped
on that machine: the SD card only has 4GB of disk space, so taking a
whole gigabyte for the bootloader seems really overkill.

Unfortunately, the kernels are not getting any smaller, so we're still
using 200MB for /boot. My /boot, with a single kernel, is currently at
90MB used, so hopefully that should be enough for two kernels in
rotation.

My /boot/efi is only 45MB here so 100MB should probably be enough.