fix default serial console configuration (tpo/tpa/team#41058)
This sample was taken from the cymru docs, where we have the peculiar
configuration of having the serial console on ttyS1 instead of the
more usual ttyS0. That said, we might need to make this customizable
as well, as it might not work everywhere.

One server at a time I guess?