just fetch the child devices directly
It seems the previous version of this code assumed the VM was running,
as it was looking for the "on primary" device and acting on that
directly. But in my tests, using real DRBD-backed VMs, that device
node would actually disappear when the VM was stopped, which is
exactly what we're doing after fecthing that information.

Now, instead, we directly fetch the child device and use that as the
primary thing. This worked during the cluster migration in
tpo/tpa/team#40972.

We also improve logging here by explicitly telling which device we're
mounting and where.