backport to cryptsetup pre-2.4
Only 2.4 has --dump-metadata-json, so we need something else. I could
have reverted to parsing the dump output itself, but previous
versions *also* seem to have a --debug-json output which we *can*
use. In my tests (in bookworm), this actually works.

As a bonus, we get access to the Version: field which is not available
in the metadata output.

I have briefly considered keeping the dump-json-metadata code and
split that in a separate function, but it seems simpler this way.